+++ CHANGELOG.md
@@ -52,7 +52,9 @@
     * ⚠️ NOTE that a misconfigured build - where the client version is not set to the same value
       as the server - would result in a false positive for an upgrade. The two should always match.
 * Calls to `Promise.track()` that are rejected with an exception will be tracked with new
-  severity level of `TrackSeverity.ERROR`
+  severity level of `TrackSeverity.ERROR`.
+* Improved client `WebSocketService` heartbeat to check that it has been receiving inbound messages
+  from the server, not just successfully sending outbound heartbeats. Attempt reconnect if not.
 
 ## v72.5.1 - 2025-04-15